+++ src/validators/AuthorizedIp.php
@@ -1,23 +1,23 @@
 <?php
-namespace selvinortiz\patrol\validators;
+namespace verbb\patrol\validators;
 
+use yii\base\Model;
 use yii\validators\Validator;
 
-/**
- * Class AuthorizedIp
- *
- * @package selvinortiz\patrol\validators
- */
-class AuthorizedIp extends Validator {
+class AuthorizedIp extends Validator
+{
+    // Public Methods
+    // =========================================================================
 
     /**
-     * @param \yii\base\Model $model
-     * @param string          $attribute
+     * @param Model $model
+     * @param string $attribute
      */
-    public function validateAttribute($model, $attribute) {
+    public function validateAttribute($model, $attribute): void
+    {
         $value = $model->{$attribute};
 
-        if (! is_array($value)) {
+        if (!is_array($value)) {
             $value = [];
         }