+++ src/variables/PatrolVariable.php
@@ -1,0 +1,15 @@
+<?php
+namespace verbb\patrol\variables;
+
+use verbb\patrol\Patrol;
+
+class PatrolVariable
+{
+    // Public Methods
+    // =========================================================================
+
+    public function getPluginName()
+    {
+        return Patrol::$plugin->getPluginName();
+    }
+}