+++ CHANGELOG.md
@@ -1,13 +1,22 @@
-# Patrol Changelog
-> Maintenance Mode & SSL Routing for [Craft 3](http://craftcms.com)
+# Changelog
+
+## 4.0.0 - 2022-07-26
+
+> {note} The plugin’s package name has changed to `verbb/patrol`. Patrol will need be updated to 4.0 from a terminal, by running `composer require verbb/patrol && composer remove selvinortiz/patrol`.
+
+### Changed
+- Migration to `verbb/patrol`.
+- Now requires Craft 3.7+.
 
 ## 3.1.3 - 2019-12-19
+
 ### Fixed
 - Fixed issue [#16] where `sslRoutingBaseUrl` was causing issue when set to `empty` or `/`
 
 [#16]: https://github.com/selvinortiz/craft-plugin-patrol/issues/16
 
 ## 3.1.2 - 2019-11-23
+
 ### Fixed
 - Fixed issue [#12] where the requesting IP could not be determined if behind some proxies
 - Fixed critical issue where Patrol would explode on sites without a proper `baseUrl` set
@@ -20,8 +29,8 @@
 
 > {note} Thank you, Simon Davies and Chris Rowe for your feedback and PRs
 
-
 ## 3.1.1 - 2019-08-25
+
 ### Fixed
 - Fixed typo in previous CHANGELOG entry
 - Fixed an issue with plugin store versioning (maybe)
@@ -30,6 +39,7 @@
 - Updated composer dependencies
 
 ## 3.1.0 - 2019-08-22
+
 ### Added
 - Added the ability to configure the `redirect status code`
 - Added the ability to enforce a primary domain
@@ -43,12 +53,15 @@
 - Updated config settings and APIs to align with Craft
 
 ## 3.0.2 - 2017-04-01
+
 ### Changed
 - Compatibility with newer Craft 3 versions
 
 ## 3.0.1 - 2017-02-03
+
 ### Fixed
 - Fixed an issue where SSL Routing was forced by default
+
 ### Changed
 - Updated file based settings support